Shut down wifi if WifiIfaceInfo is out of sync with cache
HDM relies on WifiIfaceInfo (from the chip) to be in sync with the local
cache of InterfaceCacheEntry. If this is out of sync (i.e. the chip is
reporting more ifaces than we have in the cache), then WifiIfaceInfo may
not be updated with the requestor worksource stored in the cache,
causing an NPE when we try to reference it during interface conflict
handling.

Currently we already shut down wifi if the cache has more entries than
the chip values, so we should also be doing this for the opposite case
(chip reports more ifaces than the cache).

Flag: EXEMPT minor bugfix

Bug: 352290141
Test: atest HalDeviceManagerTest
CRs-Fixed: 4036679
Change-Id: I5bc340116b80086cff98fe41e7b5b41d3ef9775f
(cherry-pick from I389d8ccf615d39f7c07022fce75fc3f1a5845e65)